Introduce ExitStatus struct
Prior to this change, the only means to customise the error message was
via the `Display` trait.

This change introduces the `ExitStatus` struct and defaults the type
parameter of `Exit` to be this struct.  Custom error types can continue
to implement `From<T> for i32` but for full control they can instead (or
in addition to) implement `From<T> for exit::ExitStatus`.

Note that implementing `From<T> for exit::ExitStatus` requires the
implementor's crate to opt-in to the unstable specialization feature.

Signed-off-by: Corin Lawson <corin@responsight.com>